Variables in python
Variables are used to store values so we can refer to them later. A variable is like a label, and you use the '=' symbol, known as the assignment operator, to assign a value to a variable. An assignment can be chained, e.g. a = b = 2 

Change the value stored in the variable greetings.